# src/loaders/logger/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Rest base repo",
    "description": "This is an API base to start new projects.",
    "version": "0.0.1",
    "contact": {
      "email": "martinfsandander@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Desarrollo"
    },
    {
      "url": "https://app-node-06-restserver.herokuapp.com/api/v1",
      "description": "Produccion"
    }
  ],
  "tags": [
    {
      "name": "users",
      "description": "services about users"
    }
  ],
  "paths": {
    "/users/": {
      "get": {
        "tags": ["users"],
        "summary": "get a list of all existing users",
        "description": "This resource get a list of users in the system.",
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "create a new user",
        "description": "Adds an new user to the system",
        "operationId": "createUser",
        "requestBody": {
          "description": "Inventory user to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "user created"
          },
          "400": {
            "description": "invalid input, object invalid"
          }          
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": ["users"],
        "summary": "user update",
        "description": "This resource allows update an user by id.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user that needs to be updated.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "delete an user",
        "description": "delete an existing user to the system by id",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user that needs to be deleted.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "invalid input, object invalid"
          }         
        }
      }
    }
  },

  "components": {
    "schemas": {
      "User": {
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "uuid",
            "example": "2537f5b3-1128-40b0-80d7-e0154611df1e"
          },
          "name": {
            "type": "string",
            "example": "martin"
          }
        }
      }
    }
  }
}